Moved the lower limit for user and group id's up to avoid conflicts with Mac OS X.

+++ src/schemas/gasharl/src/arlut/csd/ganymede/gasharl/groupCustom.java
@@ -162,7 +162,7 @@ public class groupCustom extends DBEditObject implements SchemaConstants, groupS
   public ReturnVal initializeNewObject()
   {
     ReturnVal retVal;
-    Integer gidVal = new Integer(1001);
+    Integer gidVal = new Integer(2000);
 
     /* -- */
